% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin_package}
\alias{mybin_package}
\title{Simulate a Binomial Distribution}
\usage{
mybin_package(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials per iteration}

\item{p}{Probability of success}
}
\value{
A barplot of simulated proportions and a table of frequencies
}
\description{
This function simulates repeated binomial trials and plots
the distribution of successes.
}
\examples{
mybin_package(iter = 1000, n = 10, p = 0.5)
}
